backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
pytest==8.3.4
